% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bladder1_recforest.R
\docType{data}
\name{bladder1_recforest}
\alias{bladder1_recforest}
\title{bladder1_recforest : Bladder Cancer Recurrences}
\format{
A data frame with 294 rows (118 individuals) and 8 variables:
\describe{
\item{ id }{  Patient id }
\item{ t.start }{  Start time }
\item{ t.stop }{  Stop time }
\item{ treatment }{  Placebo, pyridoxine (vitamin B6), or thiotepa }
\item{ number }{  Initial number of tumors (8=8 or more) }
\item{ size }{  Size (cm) of largest initial tumor }
\item{ death }{  Death event }
\item{ event }{  Recurrence event }
}
}
\source{
Script to generate the data can be explored using \code{browseURL(system.file("generate_bladder1_recforest.R", package = "recforest"))}
}
\usage{
bladder1_recforest
}
\description{
Preparation of the survival::bladder1 dataset for the recforest package. Please run \code{?survival::bladder1} for more information.
}
\keyword{datasets}
