% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_s3_methods.R
\name{is_supported_variable}
\alias{is_supported_variable}
\title{Check if a Variable is Supported}
\usage{
is_supported_variable(x)
}
\arguments{
\item{x}{The variable to be checked.}
}
\value{
A logical value indicating whether the variable is supported.
}
\description{
This function is a generic method that checks if a given variable is supported.
The actual implementation of the check is provided by specific methods for different classes of variables.
}
