% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_s3_methods.R
\name{make_decision}
\alias{make_decision}
\title{Make a Decision Based on Input}
\usage{
make_decision(x, value)
}
\arguments{
\item{x}{An object for which a decision needs to be made.}

\item{value}{A value that influences the decision-making process.}
}
\value{
The result of the decision-making process, which depends on the specific method implementation.
}
\description{
This function serves as a generic method for making decisions based on the input \code{x} and \code{value}.
It dispatches to the appropriate method depending on the class of \code{x}.
}
