% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_s3_methods.R
\name{plot.recforest}
\alias{plot.recforest}
\title{Plot method for recforest objects}
\usage{
\method{plot}{recforest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{recforest}.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
No return value, used for side effect.
}
\description{
Plot method for recforest objects
}
