% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recforest-package.R
\docType{package}
\name{recforest-package}
\alias{recforest}
\alias{recforest-package}
\title{recforest: Random Survival Forest for Recurrent Events}
\description{
Analyze recurrent events with right-censored data and the potential presence of a terminal event (that prevents further occurrences, like death). 'recofest' extends the random survival forest algorithm, adapting splitting rules and node estimators to handle complexities of recurrent events. The methodology is fully described in Murris, J., Bouaziz, O., Jakubczak, M., Katsahian, S., & Lavenu, A. (2024) (\url{https://hal.science/hal-04612431v1/document}).
}
\author{
\strong{Maintainer}: Juliette Murris \email{murris.juliette@gmail.com} (\href{https://orcid.org/0000-0002-7017-9865}{ORCID})

Authors:
\itemize{
  \item Guillaume Desachy \email{guillaume.desachy@pierre-fabre.com} (\href{https://orcid.org/0000-0002-2000-1436}{ORCID})
  \item Colin Fay \email{colin@thinkr.fr} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})
  \item Yohann Mansiaux \email{yohann@thinkr.fr} (\href{https://orcid.org/0000-0002-8905-6603}{ORCID})
  \item Audrey Lavenu \email{audrey.lavenu@univ-rennes.fr} (\href{https://orcid.org/0000-0002-0049-2397}{ORCID})
  \item Sandrine Katsahian \email{sandrine.katsahian@aphp.fr} (\href{https://orcid.org/0000-0002-7261-0671}{ORCID})
}

}
\keyword{internal}
