% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_s3_methods.R
\name{summary.recforest}
\alias{summary.recforest}
\title{Summary Method for recforest Objects}
\usage{
\method{summary}{recforest}(object, ...)
}
\arguments{
\item{object}{An object of class \code{recforest}.}

\item{...}{Additional arguments to be passed to the summary function.}
}
\value{
The function prints the metrics of the \code{recforest} object.
}
\description{
This function provides a summary of a recforest object by printing its metrics.
}
